requirements.txt
# Data Handling & Database
pandas
sqlalchemy
psycopg2-binary
python-dotenv

# API & Backend
fastapi
uvicorn
requests
pydantic

# AI & Semantic Search
openai

# Frontend
streamlit